Move jinja2.py to util
